Replace --no-prepare with --no-config and --init
"prepare" consists of:
a. generating a vault token
b. put the token in the correct environment variable
c. generate the terraform config from the appropriate flake attribute
d. nuke the `.terrform` directory and re-init.

we always want a and b to happen so `--no-prepare` is useless.
c is very quick thanks to flake eval caching so it's an opt-out (`--no-config`);
on the other hand, d is "expensive" (it takes quite a few seconds) and
I'm guessing that tht use-case would involve tinkering with existing
state (i.e., `terraform state`).