add BitteCluster struct
BitteCluster describes all the relevant information for the cli's
operation while dropping extraneous detail. In addition, this struct
is build from asynchronous API calls instead of the previous method of
calling the required APIs synchronously.

This is the first step toward a more data driven design for the cli
which should make the implementation simpler and the execution much
faster.