nomad token is now only used for `ssh -j`
Since it is no longer a hard requirement, move it under the `ssh`
subcommand so it doesn't pollute the cli help of other commands.

In addition, since we no longer have an on disk cache, only call
the Nomad APIs if we are actually in need of its value to save time.